エラーの表示方法を調整
